Gaussian rendering

+++ src/cuda/render_utils.h
@@ -3,7 +3,7 @@
 struct RenderUtils {
     static void *temp_storage;
     static size_t temp_storage_size;
-    static void sort_gaussians_gpu(int* keys_in, int* keys_out, float* val_in, float* val_out, int num);
+    static void sort_gaussians_gpu(float* keys_in, float* keys_out, int* val_in, int* val_out, int num);
 };
